feat: fav icon can be changed on individual entry
